Merging to updated version.

+++ 410-yardwand/app/models/Stat.java
@@ -1,6 +1,6 @@
 package models;
 
 public class Stat {
-	public int weight;
-	public int stepSpeed;
+	public Double weight;
+	public Double stepSpeed;
 }
